\name{pChen}
%\alias{pChen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Chen-weights 
}
\description{
  Estimates Chen-weights. For internal use only. Users should use the 
  wrapper \code{\link{Chenprob}}.
  
}
%\usage{
%  summary.wpl(object,...)  {
%  }
  %- maybe also 'usage' for other objects documented here.
%  \arguments{
%    \item{object}{
%      the result of a wpl fit
%    }
%    \item{...}{
%      for future methods
%    }
%  }
  
  %\value{
    %  The calling argument is returned.
    %}
  
  \author{
    Nathalie C. Stoer
  }
  
  %% ~Make other sections like Warning with \section{Warning }{....} ~
    
    \seealso{
      \code{\link{wpl}}, \code{\link{Chenprob}}
    }
  
  % Add one or more standard keywords, see file 'KEYWORDS' in the
  % R documentation directory.
  %\keyword{ ~kwd1 }
  %\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
  