\name{pGAM}
\alias{pGAM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Generlaized additive model weights 
}
\description{
  Estimates GAM-weights. For internal use only. Users should use the 
  wrapper \code{\link{GAMprob}}.
  
}
%\usage{
  %  summary.wpl(object,...)  {
    %  }
  %- maybe also 'usage' for other objects documented here.
  %  \arguments{
    %    \item{object}{
      %      the result of a wpl fit
      %    }
    %    \item{...}{
      %      for future methods
      %    }
    %  }
  
  %\value{
    %  The calling argument is returned.
    %}
  
  \author{
    Nathalie C. Stoer
  }
  
  %% ~Make other sections like Warning with \section{Warning }{....} ~
    
    \seealso{
      \code{\link{wpl}}, \code{\link{GAMprob}}, \code{\link[gam]{gam}}
    }
  
  % Add one or more standard keywords, see file 'KEYWORDS' in the
  % R documentation directory.
  %\keyword{ ~kwd1 }
  %\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
  