\name{print.wpl}
\alias{print.wpl}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Print a wpl object
}
\description{
  Prints the fit of (each) weighted Cox-regression
}
\usage{
  \method{print}{wpl}(x,...)  
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  The result of a call to wpl
  }
  \item{...}{
    For future methods
  }
}

%\value{
%  The calling argument is returned.
%}

\author{
  Nathalie C. Stoer
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
  
  \seealso{
    \code{\link{wpl}}
  }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
