\name{multipleNCC-internal}
\alias{multipleNCC-internal}
%\alias{ModelbasedVar}
%\alias{PoststratVar}
%\alias{print.wpl}
%\alias{summary.wpl}
\alias{wpl.default}
\alias{wpl.formula}
\alias{wplEst}
%\alias{pChen}
%\alias{pGAM}
%\alias{pGLM}
%\alias{pKM}
\docType{package}
\title{
Internal function
}
\description{
Internal function
}

\author{
Nathalie C. Stoer
}



